Refactoring router
Refactored the route to be used with only bare minimum information (size of map, where to start, where to go, and a way to know whether a move is possible and how much that move costs). Next would be Javadoc'ing the class, and creating seperate test classes and test data. Due to smaller size of the classes involved this should be easier now)